working buzzer and blinking LED
