+++ AvionicsSoftware.ioc
@@ -111,39 +111,38 @@ Mcu.Pin2=PH0-OSC_IN
 Mcu.Pin20=PB13
 Mcu.Pin21=PB14
 Mcu.Pin22=PB15
-Mcu.Pin23=PC6
-Mcu.Pin24=PC7
-Mcu.Pin25=PC8
-Mcu.Pin26=PC9
-Mcu.Pin27=PA8
-Mcu.Pin28=PA9
-Mcu.Pin29=PA10
+Mcu.Pin23=PC7
+Mcu.Pin24=PC8
+Mcu.Pin25=PC9
+Mcu.Pin26=PA8
+Mcu.Pin27=PA9
+Mcu.Pin28=PA10
+Mcu.Pin29=PA11
 Mcu.Pin3=PH1-OSC_OUT
-Mcu.Pin30=PA11
-Mcu.Pin31=PA13
-Mcu.Pin32=PA14
-Mcu.Pin33=PA15
-Mcu.Pin34=PC10
-Mcu.Pin35=PC11
-Mcu.Pin36=PC12
-Mcu.Pin37=PD2
-Mcu.Pin38=PB3
-Mcu.Pin39=PB4
+Mcu.Pin30=PA13
+Mcu.Pin31=PA14
+Mcu.Pin32=PA15
+Mcu.Pin33=PC10
+Mcu.Pin34=PC11
+Mcu.Pin35=PC12
+Mcu.Pin36=PD2
+Mcu.Pin37=PB3
+Mcu.Pin38=PB4
+Mcu.Pin39=PB5
 Mcu.Pin4=PC0
-Mcu.Pin40=PB5
-Mcu.Pin41=PB6
-Mcu.Pin42=PB7
-Mcu.Pin43=VP_CRC_VS_CRC
-Mcu.Pin44=VP_FREERTOS_VS_CMSIS_V1
-Mcu.Pin45=VP_SYS_VS_tim1
-Mcu.Pin46=VP_TIM2_VS_ClockSourceINT
-Mcu.Pin47=VP_TIM3_VS_ClockSourceINT
+Mcu.Pin40=PB6
+Mcu.Pin41=PB7
+Mcu.Pin42=VP_CRC_VS_CRC
+Mcu.Pin43=VP_FREERTOS_VS_CMSIS_V1
+Mcu.Pin44=VP_SYS_VS_tim1
+Mcu.Pin45=VP_TIM2_VS_ClockSourceINT
+Mcu.Pin46=VP_TIM3_VS_ClockSourceINT
 Mcu.Pin5=PC2
 Mcu.Pin6=PC3
 Mcu.Pin7=PA0-WKUP
 Mcu.Pin8=PA1
 Mcu.Pin9=PA2
-Mcu.PinsNb=48
+Mcu.PinsNb=47
 Mcu.ThirdPartyNb=0
 Mcu.UserConstants=
 Mcu.UserName=STM32F405RGTx
@@ -198,7 +197,7 @@ PA14.Locked=true
 PA14.Mode=Trace_Asynchronous_SW
 PA14.Signal=SYS_JTCK-SWCLK
 PA15.GPIOParameters=GPIO_Label
-PA15.GPIO_Label=BUZZER_PWM
+PA15.GPIO_Label=HDI_PWM_BUZZER
 PA15.Locked=true
 PA15.Signal=S_TIM2_CH1_ETR
 PA2.GPIOParameters=GPIO_Label
@@ -335,10 +334,6 @@ PC4.GPIOParameters=GPIO_Label
 PC4.GPIO_Label=IMU_MAG_CS
 PC4.Locked=true
 PC4.Signal=GPIO_Output
-PC6.GPIOParameters=GPIO_Label
-PC6.GPIO_Label=HDI_PWM_BUZZER
-PC6.Locked=true
-PC6.Signal=S_TIM3_CH1
 PC7.GPIOParameters=GPIO_Label
 PC7.GPIO_Label=LED_3
 PC7.Locked=true
@@ -434,8 +429,6 @@ SH.GPXTI15.0=GPIO_EXTI15
 SH.GPXTI15.ConfNb=1
 SH.S_TIM2_CH1_ETR.0=TIM2_CH1,PWM Generation1 CH1
 SH.S_TIM2_CH1_ETR.ConfNb=1
-SH.S_TIM3_CH1.0=TIM3_CH1,PWM Generation1 CH1
-SH.S_TIM3_CH1.ConfNb=1
 SPI1.BaudRatePrescaler=SPI_BAUDRATEPRESCALER_256
 SPI1.CLKPhase=SPI_PHASE_2EDGE
 SPI1.CLKPolarity=SPI_POLARITY_HIGH
@@ -458,13 +451,14 @@ SPI3.Mode=SPI_MODE_MASTER
 SPI3.VirtualType=VM_MASTER
 TIM2.AutoReloadPreload=TIM_AUTORELOAD_PRELOAD_ENABLE
 TIM2.Channel-PWM\ Generation1\ CH1=TIM_CHANNEL_1
-TIM2.ClockDivision=TIM_CLOCKDIVISION_DIV1
-TIM2.IPParameters=Channel-PWM Generation1 CH1,AutoReloadPreload,Period,Prescaler,ClockDivision,Pulse-PWM Generation1 CH1
+TIM2.IPParameters=Prescaler,Period,AutoReloadPreload,Channel-PWM Generation1 CH1,Pulse-PWM Generation1 CH1
 TIM2.Period=210-1
 TIM2.Prescaler=952-1
 TIM2.Pulse-PWM\ Generation1\ CH1=50
-TIM3.Channel-PWM\ Generation1\ CH1=TIM_CHANNEL_1
-TIM3.IPParameters=Channel-PWM Generation1 CH1
+TIM3.AutoReloadPreload=TIM_AUTORELOAD_PRELOAD_ENABLE
+TIM3.IPParameters=Prescaler,Period,AutoReloadPreload
+TIM3.Period=210-1
+TIM3.Prescaler=952-1
 UART4.BaudRate=38400
 UART4.IPParameters=VirtualMode,BaudRate
 UART4.VirtualMode=Asynchronous
