+++ Components/FlightControl/FlightTask.cpp
@@ -67,7 +67,8 @@ void FlightTask::Run(void * pvParams)
 
 
         //Every cycle, print something out (for testing)
-        SOAR_PRINT("FlightTask::Run() - [%d] Seconds\n", tempSecondCounter++);
+    	SOAR_PRINT("FlightTask::Run() - [%d] Seconds\n", tempSecondCounter++);
+		osDelay(1000);
 
         //Process any commands, in non-blocking mode (TODO: Change to instant-processing once complete HID/DisplayTask)
         Command cm;
